+++ tests/access-control.tests.ts
@@ -1,142 +1,10 @@
-import { consts, Roles } from './constants'
+import { Roles } from './constants'
 import { bnArg, expect, fromSigner, setupContract } from './helpers'
 
 describe('MY_ACCESS_CONTROL', () => {
   async function setup() {
     return setupContract('my_access_control', 'new')
   }
-
-  async function setup_receiver() {
-    return setupContract('psp721_receiver', 'new')
-  }
-
-  it('PSP 721 - mint works', async () => {
-    const { contract, query, defaultSigner: sender } = await setup()
-
-    // Arrange - Ensure sender balance is 0
-    await expect(query.balanceOf(sender.address)).to.have.output(0)
-
-    // Act - Sender mint a token
-    await expect(contract.tx.mint(bnArg(0))).to.eventually.be.fulfilled
-
-    // Assert - Sender balance is now 1
-    await expect(query.balanceOf(sender.address)).to.have.output(1)
-  })
-
-  it('PSP 721 - mint existing should fail', async () => {
-    const { contract, query, defaultSigner: sender } = await setup()
-
-    // Arrange - Sender mint a token with id [0; 32] and have balance of 1
-    await expect(contract.tx.mint(bnArg(0))).to.eventually.be.fulfilled
-    await expect(query.balanceOf(sender.address)).to.have.output(1)
-    await expect(query.ownerOf(bnArg(0))).to.have.output(sender.address)
-
-    // Act - Minting the id [0; 32] should fail
-    await expect(contract.tx.mint(bnArg(0))).to.eventually.be.rejected
-
-    // Assert - Alice balance is still 1
-    await expect(query.balanceOf(sender.address)).to.have.output(1)
-  })
-
-  it('PSP 721 - approved transfer works', async () => {
-    const {
-      contract,
-      query,
-      defaultSigner: sender,
-      accounts: [alice, bob]
-    } = await setup()
-
-    // Arrange - Sender mint a Token and Approve Alice as spender of this token
-    await expect(contract.tx.mint(bnArg(0))).to.eventually.be.fulfilled
-    await expect(query.ownerOf(bnArg(0))).to.have.output(sender.address)
-    await expect(contract.tx.approve(alice.address, bnArg(0))).to.eventually.be.fulfilled
-
-    // Act - Alice transfer the token form sender to bob
-    await expect(fromSigner(contract, alice.address).tx.transferFrom(sender.address, bob.address, bnArg(0))).to.eventually.be.fulfilled
-
-    // Assert - Bob is now owner of the token
-    await expect(query.ownerOf(bnArg(0))).to.have.output(bob.address)
-  })
-
-  it('PSP 721 - safe transfer works', async () => {
-    const {
-      tx,
-      query,
-      defaultSigner: sender,
-      accounts: [alice]
-    } = await setup()
-
-    const { contract } = await setup_receiver()
-
-    // Arrange - Sender mint a Token and Approve Receiver as spender of this token
-    await expect(tx.mint(bnArg(0))).to.eventually.be.fulfilled
-    await expect(query.ownerOf(bnArg(0))).to.have.output(sender.address)
-
-    // Act - Alice transfer the token form sender to bob
-    await expect(contract.query.getCallCounter()).to.have.output(0)
-    await expect(tx.safeTransferFrom(sender.address, contract.address, bnArg(0), 'data')).to.eventually.be.fulfilled
-    await expect(contract.query.getCallCounter()).to.have.output(1)
-
-    // Assert - Bob is now owner of the token
-    await expect(query.ownerOf(bnArg(0))).to.have.output(contract.address.toString())
-  })
-
-  it('PSP 721 - receiver can reject the transfer', async () => {
-    const { tx, query, defaultSigner: sender } = await setup()
-
-    const { contract } = await setup_receiver()
-
-    // Arrange - Sender mint a token
-    await expect(tx.mint(bnArg(0))).to.eventually.be.fulfilled
-    await expect(query.ownerOf(bnArg(0))).to.have.output(sender.address)
-
-    // Act - Receiver wants to reject the next transfer
-    await expect(contract.tx.revertNextTransfer()).to.eventually.be.fulfilled
-
-    // Assert - Sender cannot send token to receiver & Sender still own the token
-    await expect(tx.safeTransferFrom(sender.address, contract.address, bnArg(0), 'data')).to.eventually.be.rejected
-    await expect(query.ownerOf(bnArg(0))).to.have.output(sender.address)
-  })
-
-  it('PSP 721 - approved for all works', async () => {
-    const {
-      contract,
-      query,
-      defaultSigner: sender,
-      accounts: [alice, bob]
-    } = await setup()
-
-    // Arrange - Sender mint 2 tokens and Approve Alice to spend on all his tokens
-    await expect(contract.tx.mint(bnArg(0))).to.eventually.be.fulfilled
-    await expect(contract.tx.mint(bnArg(1))).to.eventually.be.fulfilled
-    await expect(query.balanceOf(sender.address)).to.have.output(2)
-    await expect(contract.tx.setApprovalForAll(alice.address, true)).to.eventually.be.fulfilled
-
-    // Act - Alice Transfer the two tokens to bob
-    await expect(fromSigner(contract, alice.address).tx.transferFrom(sender.address, bob.address, bnArg(0))).to.eventually.be.fulfilled
-    await expect(fromSigner(contract, alice.address).tx.transferFrom(sender.address, bob.address, bnArg(1))).to.eventually.be.fulfilled
-
-    // Assert - Bob owns the two tokens
-    await expect(query.ownerOf(bnArg(0))).to.have.output(bob.address)
-    await expect(query.ownerOf(bnArg(1))).to.have.output(bob.address)
-
-    // Assert - Bob is now owner of the tokens:c check that sender cannot transfer tokens on his behalf
-    await expect(fromSigner(contract, sender.address).tx.transferFrom(bob.address, alice.address, bnArg(0))).to.eventually.be.rejected
-  })
-
-  it('PSP 721 - burn works', async () => {
-    const { contract, query, defaultSigner: sender } = await setup()
-
-    // Arrange - Alice mint a token & have a balance of 1
-    await expect(contract.tx.mint(bnArg(0))).to.eventually.be.fulfilled
-    await expect(query.balanceOf(sender.address)).to.have.output(1)
-
-    // Act - Alice burn the token
-    await expect(contract.tx.burn(bnArg(0))).to.eventually.be.fulfilled
-
-    // Assert - Alice balance is 0
-    await expect(query.balanceOf(sender.address)).to.have.output(0)
-  })
 
   it('ACCESS CONTROL - only minter role is allowed to mint', async () => {
     const {
