+++ tests/ownable.tests.ts
@@ -6,117 +6,6 @@ describe('MY_OWNABLE', () => {
     return setupContract('my_ownable', 'new')
   }
 
-  async function setup_receiver() {
-    return setupContract('psp1155_receiver', 'new')
-  }
-
-  it('PSP 1155 - mint works', async () => {
-    const { contract, query, defaultSigner: sender } = await setup()
-
-    // Arrange - Ensure sender balance is 0
-    await expect(query.balanceOf(sender.address, bnArg(0))).to.have.output(0)
-
-    // Act - Sender mint a token
-    await expect(contract.tx.mint(bnArg(0), 1)).to.eventually.be.fulfilled
-
-    // Assert - Sender balance is now 1
-    await expect(query.balanceOf(sender.address, bnArg(0))).to.have.output(1)
-  })
-
-  it('PSP 1155 - burn works', async () => {
-    const { contract, query, defaultSigner: sender } = await setup()
-
-    // Arrange - Sender mint a token & have a balance of 1
-    await expect(contract.tx.mint(bnArg(0), 1)).to.eventually.be.fulfilled
-    await expect(query.balanceOf(sender.address, bnArg(0))).to.have.output(1)
-
-    // Act - Sender burn the token
-    await expect(contract.tx.burn(bnArg(0), 1)).to.eventually.be.fulfilled
-
-    // Assert - Sender balance is 0
-    await expect(query.balanceOf(sender.address, bnArg(0))).to.have.output(0)
-  })
-
-  it('PSP 1155 - transfer works', async () => {
-    const {
-      contract,
-      query,
-      defaultSigner: sender,
-      accounts: [alice]
-    } = await setup()
-
-    // Arrange - Sender mint 100 token
-    await expect(contract.tx.mint(bnArg(0), 100)).to.eventually.be.fulfilled
-
-    // Act - Sender transfer 50 token to Alice
-    await expect(contract.tx.safeTransferFrom(sender.address, alice.address, bnArg(0), 50, 'data')).to.eventually.be.fulfilled
-
-    // Assert - Alice own 50 token
-    await expect(query.balanceOf(alice.address, bnArg(0))).to.have.output(50)
-  })
-
-  it('PSP 1155 - safe batch transfer works', async () => {
-    const { tx, query, defaultSigner: sender } = await setup()
-
-    const { contract } = await setup_receiver()
-
-    // Arrange - Sender mint 100 of token 0 and 100 of token 1
-    await expect(tx.mint(bnArg(0), 100)).to.eventually.be.fulfilled
-    await expect(tx.mint(bnArg(1), 100)).to.eventually.be.fulfilled
-
-    // Act - Sender transfer 20 of token 0 and 70 of token 1
-    await expect(contract.query.getCallCounter()).to.have.output(0)
-    await expect(tx.safeBatchTransferFrom(sender.address, contract.address, [[bnArg(0), 20], [bnArg(1), 70]], 'data')).to.eventually.be.fulfilled
-    await expect(contract.query.getCallCounter()).to.have.output(1)
-
-    // Assert - Receiver own 20 of token 0 and own 70 of token 1
-    await expect(
-      query.balanceOfBatch([[contract.address, bnArg(0)], [sender.address, bnArg(0)], [contract.address, bnArg(1)], [sender.address, bnArg(1)]])
-    ).to.have.output([20, 80, 70, 30])
-  })
-
-  it('PSP 1155 - receiver can reject the transfer', async () => {
-    const { tx, query, defaultSigner: sender } = await setup()
-
-    const { contract } = await setup_receiver()
-
-    // Arrange - Sender mint a token
-    await expect(tx.mint(bnArg(0), 1)).to.eventually.be.fulfilled
-
-    // Act - Receiver wants to reject the next transfer
-    await expect(contract.tx.revertNextTransfer()).to.eventually.be.fulfilled
-
-    // Assert - Sender cannot send token to receiver
-    await expect(tx.safeTransferFrom(sender.address, contract.address, bnArg(0), 1, 'data')).to.eventually.be.rejected
-    await expect(query.balanceOfBatch([[contract.address, bnArg(0)], [sender.address, bnArg(0)]])).to.have.output([0, 1])
-  })
-
-  it('PSP 1155 - approved for all works', async () => {
-    const {
-      contract,
-      query,
-      defaultSigner: sender,
-      accounts: [alice, bob]
-    } = await setup()
-
-    // Arrange - Sender mint 2 tokens and Approve Alice to spend on all his tokens
-    await expect(contract.tx.mint(bnArg(0), 1)).to.eventually.be.fulfilled
-    await expect(contract.tx.mint(bnArg(1), 1)).to.eventually.be.fulfilled
-    await expect(query.balanceOfBatch([[sender.address, bnArg(0)], [sender.address, bnArg(1)]])).to.have.output([1, 1])
-    await expect(contract.tx.setApprovalForAll(alice.address, true)).to.eventually.be.fulfilled
-
-    // Act - Alice Transfer the two tokens to bob
-    await expect(fromSigner(contract, alice.address).tx.safeBatchTransferFrom(sender.address, bob.address, [[bnArg(0), 1], [bnArg(1), 1]], 'data')).to
-      .eventually.be.fulfilled
-
-    // Assert - Bob owns the two tokens+
-    await expect(query.balanceOfBatch([[bob.address, bnArg(0)], [bob.address, bnArg(1)]])).to.have.output([1, 1])
-
-    // Assert - Bob is now owner of the tokens: check that sender cannot transfer tokens on his behalf
-    await expect(fromSigner(contract, alice.address).tx.safeBatchTransferFrom(bob.address, alice.address, [[bnArg(0), 1], [bnArg(1), 1]], 'data')).to
-      .eventually.be.rejected
-  })
-
   it('OWNABLE - owner is by default contract deployer', async () => {
     const { query, defaultSigner: sender } = await setup()
